Feature/kaleb coberly/user requests (#21)
Adds user requests:
- Move all designated book-one drivers to the first split workbook.
(And, designate with constant or file.) The client actually retracted
this request, but I'd already implemented it. It doesn't do anything
unless you supply the file or update the constant, so I'm leaving it.
- Adds extra notes as a footnote. This was a nice-to-have. We'll see if
they like it where I put it, in the notes column to take advantage of
the already set width and allow word wrapping. (Can't seem to word wrap
merged cells.) (ExtraNotes is a class with a notes member, instead of
just a constant data frame or just a data frame member, so it will show
up in docs.)
- Defaults manifest date to the next Friday.
- Makes driver names title case and leaves recipient name casing alone
(instead of to upper).
- Moves drivers with same name but different numbers to the same
workbook.


Also:
- Adds See Also sections to docs.
- Moves sphinxcontrib-mermaid to docs dependencies.
- Renames validation_df to formatting_df to follow convention that
transformations can validate, but validations should never transform.