Debug/kaleb coberly/debug black (#136)
Was an attempt to fix a problem with `black` that didn't exist, thought
to cause local formatting and lint checking to function differently than
lint checking on GitHub runner. The issue was not `black` but that the
`pull_request_workflow` triggers in the `shared` repo were checking out
the target branch's source code and thus running QC/tests on it and
missing the changes in the PR branch.

We recently replaced the `pull_request` workflow triggers with
`pull_request_target` to make sure that the target-branch state of the
workflows run on a PR, instead of running workflows with whatever
changes the PR included to them. The issue was fixed in `shared`
reusable workflows, passing the PR ref to the checkout action, allowing
the target-branch workflow to still run while checking out the PR source
code to run QC on.

Due to the process of debugging, this PR reformats the source code.
- Most of the reformats are git renormalization of line endings to lf.
- Some are due to temporarily using `preview = true` in the `shared`
`black` config, causing forward-looking formats to be applied that were
then not reverted when the config was reset.
- There also may be some formatting due to removing the deprecated
`black` config `skip_trailing_commas = true` (in the `shared` config).

There's also a new line added between a comment block and a class def at
the top of `api_callers.py`. This was the original formatting change
that triggered the debug in the first place (due to not being
incorporated in the target branch when QC was run on it).