Bugfix/kaleb coberly/no address name (#168)
Drops `addressName` in stop uploads.

We were passing the whole address in as `addressName` out of confusion
about what Circuit was doing with the fields. The post fields include
`addressName`, `addressLineOne`, `addressLineTwo`, `city`, `state`,
`zip` ... But, the returned fields are only `address`, `addressLineOne`,
and `addressLineTwo`. Circuit adds `addressName` to the beginning of
`addressLineOne`, then does some deduplication. So, we didn't notice a
problem until they changed the deduplication so that it stops
deduplicating.

Removing `addressName` altogether, since we don't pass the recipient
names to the address, fixes the problem.

That said, we may need to change the way we create `addressLineTwo` in
the future if they stop deduplicating that too. We currently put
city/state/zip in `addressLineTwo`. If we stop doing that and only pass
apartment numbers to `addressLineTwo`, then Circuit will add to the end
of `addressLineTwo` the city/state/zip that we passed in to those
specific fields. Saving that for another PR since this needs to go into
prod for tomorrow.

(Also makes a Mermaid chart easier to read in the markdown. Unrelated
porkbarrel.)